requirements.txt
langchain
langchain-huggingface
langchain-community
python-dotenv
faiss-cpu
sentence-transformers
pypdf